requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==3.6.4.0
boto3==1.28.68
botocore==1.31.68
celery==5.2.7
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography>=41.0.5,<45
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.8
django-celery-results==2.5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.3
django-ses==3.5.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.4
jmespath==1.0.1
kombu==5.3.5
Markdown==3.5
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
python-magic==0.4.27
python-magic-bin==0.4.14
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
sentry-sdk==2.21.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
